fix: rgb issue

+++ src/style.rs
@@ -353,7 +353,7 @@ impl FromStr for Color {
                     u8::from_str_radix(&s[3..5], 16),
                     u8::from_str_radix(&s[5..7], 16),
                 ) {
-                    Self::Rgb(r as u8, g as u8, b as u8)
+                    Self::Rgb(r, g, b)
                 } else if let Ok(index) = s.parse::<u8>() {
                     Self::Indexed(index)
                 } else {
